# backend/insomnia/strobe-openapi.yaml
openapi: 3.0.3
info:
  title: Strobe Local API
  version: 1.0.0
  contact:
    name: CAB432 Teaching Team
    email: cab432@qut.edu.au
  description: |
    Local OpenAPI specification for the Strobe server.

    This spec is designed for direct import into Insomnia or other API clients and reflects
    the request and response shapes implemented by the Express server in `server/src`.

    Authentication uses JWT bearer tokens. Register or log in first, then send:

    `Authorization: Bearer <token>`
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JWT token only. Do not include the "Bearer " prefix.
  parameters:
...
    UserIdParam:
      in: path
      name: userId
      required: true
      description: User identifier.
      schema:
        type: string
        example: user_12345678
    PostIdParam:
      in: path
      name: postId
      required: true
      description: Post identifier.
      schema:
        type: string
        example: post_12345678
    CommentIdParam:
      in: path
      name: commentId
      required: true
      description: Comment identifier.
      schema:
        type: string
        example: comment_12345678
...
    QueryLimit:
      in: query
      name: limit
      required: false
      description: Maximum number of records to return. The server caps this at 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    QueryOffset:
      in: query
      name: offset
      required: false
      description: Number of records to skip before returning results.
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Title is required
...
    UserStats:
      type: object
      required: [followers, following, posts]
      properties:
        followers:
          type: integer
          example: 12
        following:
          type: integer
          example: 7
        posts:
          type: integer
          example: 4
    User:
      type: object
      required: [id, username, email, role, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake_bradford
        email:
          type: string
          format: email
          example: jake.bradford@example.com
        role:
          type: string
          enum: [user, moderator]
          example: user
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:00:00.000Z
    EnrichedUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [stats, currentUserFollows]
          properties:
            stats:
              $ref: "#/components/schemas/UserStats"
            currentUserFollows:
              type: boolean
              example: false
    PostAuthor:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake_bradford
    PostStats:
      type: object
      required: [likes, comments]
      properties:
        likes:
          type: integer
          example: 3
        comments:
          type: integer
          example: 1
    Post:
      type: object
      required:
        [
          id,
          userId,
          title,
          description,
          images,
          status,
          createdAt,
          updatedAt,
          author,
          stats,
          currentUserLiked,
        ]
      properties:
        id:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          example: Sunrise ride
        description:
          type: string
          example: Cold start, good coffee, and an empty river path.
        images:
          type: array
          items:
            type: string
          example:
            - /uploads/550e8400-e29b-41d4-a716-446655440000/660e8400-e29b-41d4-a716-446655440001/5f9439f1.jpg
        status:
          type: string
          enum: [active, hidden]
          example: active
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:10:00.000Z
        hiddenBy:
          type: string
          nullable: true
          example: null
        hiddenAt:
          type: string
          format: date-time
          nullable: true
          example: null
        author:
          $ref: "#/components/schemas/PostAuthor"
        stats:
          $ref: "#/components/schemas/PostStats"
        currentUserLiked:
          type: boolean
          example: false
    CommentAuthor:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: jake_bradford
    Comment:
      type: object
      required: [id, postId, userId, text, createdAt, updatedAt, author]
      properties:
        id:
          type: string
          example: 770e8400-e29b-41d4-a716-446655440002
        postId:
          type: string
          example: 660e8400-e29b-41d4-a716-446655440001
        userId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        text:
          type: string
          example: Great light in that first shot.
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:16:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2026-07-08T09:16:00.000Z
        author:
          $ref: "#/components/schemas/CommentAuthor"
    Follow:
      type: object
      required: [id, followerId, followeeId, createdAt]
      properties:
        id:
          type: string
          example: 880e8400-e29b-41d4-a716-446655440003
        followerId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440010
        followeeId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        createdAt:
          type: string
          format: date-time
          example: 2026-07-08T09:20:00.000Z
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jake.bradford@example.com
        password:
          type: string
          minLength: 6
          example: password123
        role:
          type: string
          enum: [user, moderator]
          default: user
          example: user
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: jake.bradford@example.com
        password:
          type: string
          example: password123
...
    CreateCommentRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          example: Great light in that first shot.
...
    AuthResponse:
      type: object
      required: [message, user, token]
      properties:
        message:
          type: string
          example: Created successfully
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
    PostListResponse:
      type: object
      required: [message, posts]
      properties:
        message:
          type: string
          example: Success
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
    CommentResponse:
      type: object
      required: [message, comment]
      properties:
        message:
          type: string
          example: Created successfully
        comment:
          $ref: "#/components/schemas/Comment"
    CommentListResponse:
      type: object
      required: [message, comments]
      properties:
        message:
          type: string
          example: Success
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    FollowResponse:
      type: object
      required: [message, follow]
      properties:
        message:
          type: string
          example: User followed
        follow:
          $ref: "#/components/schemas/Follow"
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Deleted successfully
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: ValidationError
            message: Title is required
    Unauthorised:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorised
            message: Invalid or expired token
...
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NotFound
            message: Post not found
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register user
      description: Creates a new user account and immediately returns a JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: jake.bradford@example.com
              password: password123
              role: user
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                message: Created successfully
                user:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  username: jake_bradford
                  email: jake.bradford@example.com
                  role: user
                  createdAt: 2026-07-08T09:00:00.000Z
                  updatedAt: 2026-07-08T09:00:00.000Z
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email is already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Conflict
                message: Email already exists
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Login user
      description: Authenticates an existing user and returns a fresh JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: jake.bradford@example.com
              password: password123
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                message: Success
                user:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  username: jake_bradford
                  email: jake.bradford@example.com
                  role: user
                  createdAt: 2026-07-08T09:00:00.000Z
                  updatedAt: 2026-07-08T09:00:00.000Z
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Unauthorised
                message: Invalid email or password
...
  /v1/users/{userId}/follow:
    post:
      tags: [Follows]
      operationId: followUser
      summary: Follow user
      description: Creates a follower relationship from the authenticated user to the target user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "201":
          description: Follow created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
              example:
                message: User followed
                follow:
                  id: 880e8400-e29b-41d4-a716-446655440003
                  followerId: 550e8400-e29b-41d4-a716-446655440010
                  followeeId: 550e8400-e29b-41d4-a716-446655440000
                  createdAt: 2026-07-08T09:20:00.000Z
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: Attempted to follow yourself.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: You cannot follow yourself
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Authenticated user already follows the target user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Conflict
                message: You are already following this user
    delete:
      tags: [Follows]
      operationId: unfollowUser
      summary: Unfollow user
      description: Removes a follower relationship.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: Follow removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: User unfollowed
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Target user not found or relationship did not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                userMissing:
                  value:
                    error: NotFound
                    message: User not found
                relationshipMissing:
                  value:
                    error: NotFound
                    message: You are not following this user
  /v1/users/{userId}/followers:
    get:
      tags: [Follows]
      operationId: getFollowers
      summary: Get followers
      description: Lists enriched user objects for everyone following the target user.
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: Followers returned.
          content:
            application/json:
              schema:
                type: object
                required: [message, followers]
                properties:
                  message:
                    type: string
                  followers:
                    type: array
                    items:
                      $ref: "#/components/schemas/EnrichedUser"
              example:
                message: Success
                followers:
                  - id: 550e8400-e29b-41d4-a716-446655440010
                    username: matthew_mckague
                    email: matthew.mckague@example.com
                    role: user
                    createdAt: 2026-07-01T09:00:00.000Z
                    updatedAt: 2026-07-01T09:00:00.000Z
                    stats:
                      followers: 3
                      following: 11
                      posts: 2
                    currentUserFollows: false
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/users/{userId}/following:
    get:
      tags: [Follows]
      operationId: getFollowing
      summary: Get following
      description: Lists enriched user objects for everyone the target user currently follows.
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: Following list returned.
          content:
            application/json:
              schema:
                type: object
                required: [message, following]
                properties:
                  message:
                    type: string
                  following:
                    type: array
                    items:
                      $ref: "#/components/schemas/EnrichedUser"
              example:
                message: Success
                following:
                  - id: 550e8400-e29b-41d4-a716-446655440000
                    username: alan_yu
                    email: alan.yu@example.com
                    role: user
                    createdAt: 2026-07-08T09:00:00.000Z
                    updatedAt: 2026-07-08T09:00:00.000Z
                    stats:
                      followers: 12
                      following: 7
                      posts: 4
                    currentUserFollows: true
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/posts/{postId}/comments:
    get:
      tags: [Comments]
      operationId: getPostComments
      summary: Get comments for post
      description: Returns comments in ascending creation order.
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      responses:
        "200":
          description: Comments returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
              example:
                message: Success
                comments:
                  - id: 770e8400-e29b-41d4-a716-446655440002
                    postId: 660e8400-e29b-41d4-a716-446655440001
                    userId: 550e8400-e29b-41d4-a716-446655440010
                    text: Great light in that first shot.
                    createdAt: 2026-07-08T09:16:00.000Z
                    updatedAt: 2026-07-08T09:16:00.000Z
                    author:
                      id: 550e8400-e29b-41d4-a716-446655440010
                      username: matthew_mckague
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Comments]
      operationId: createComment
      summary: Create comment
      description: Adds a comment to a post as the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
            example:
              text: Great light in that first shot.
      responses:
        "201":
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
              example:
                message: Created successfully
                comment:
                  id: 770e8400-e29b-41d4-a716-446655440002
                  postId: 660e8400-e29b-41d4-a716-446655440001
                  userId: 550e8400-e29b-41d4-a716-446655440010
                  text: Great light in that first shot.
                  createdAt: 2026-07-08T09:16:00.000Z
                  updatedAt: 2026-07-08T09:16:00.000Z
                  author:
                    id: 550e8400-e29b-41d4-a716-446655440010
                    username: matthew_mckague
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/posts/{postId}/comments/{commentId}:
    delete:
      tags: [Comments]
      operationId: deleteComment
      summary: Delete comment
      description: Deletes a comment. Only the comment author can delete it.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
        - $ref: "#/components/parameters/CommentIdParam"
      responses:
        "200":
          description: Comment deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              example:
                message: Deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorised"
        "403":
          description: Non-author attempted delete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Forbidden
                message: You can only delete your own comments
        "404":
          description: Comment not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: Comment not found
  /v1/feed:
    get:
      tags: [Feed]
      operationId: getPersonalFeed
      summary: Get personal feed
      description: Returns posts from accounts the authenticated user follows, newest first.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryOffset"
      responses:
        "200":
          description: Feed returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
              example:
                message: Success
                posts:
                  - id: 660e8400-e29b-41d4-a716-446655440011
                    userId: 550e8400-e29b-41d4-a716-446655440010
                    title: Campus coffee stop
                    description: Fast break before the next lab.
                    images: []
                    status: active
                    createdAt: 2026-07-08T09:12:00.000Z
                    updatedAt: 2026-07-08T09:12:00.000Z
                    author:
                      id: 550e8400-e29b-41d4-a716-446655440010
                      username: matthew_mckague
                    stats:
                      likes: 1
                      comments: 0
                    currentUserLiked: false
        "401":
          $ref: "#/components/responses/Unauthorised"
        "404":
          description: Authenticated user ID no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: NotFound
                message: User not found
...